requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
supabase==2.0.0
reportlab==4.0.4
python-docx==0.8.11
python-dotenv==1.0.0
faker==19.6.2
